requirements.txt
Flask==2.0.1
easyocr==1.7.1
Pillow==9.5.0
python-dotenv==0.19.0
opencv-python==4.8.0.74
numpy==1.24.3
pdf2image==1.16.3
google-generativeai==0.3.2
gunicorn==20.1.0
flask-cors==4.0.0
